requirements.txt
# Core Framework
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0

# AI and Machine Learning
openai==1.57.0
anthropic==0.40.0
google-generativeai==0.8.3
langchain==0.3.9
langchain-openai==0.2.9
langchain-community==0.3.9
langchain-chroma==0.1.4
sentence-transformers==3.3.1
chromadb==0.5.23
pypdf2==3.0.1
python-docx==1.1.2
pytesseract==0.3.13
Pillow==10.4.0

# Communications
africastalking==1.2.9
twilio==9.3.7
requests==2.32.4
python-telegram-bot==21.9

# Voice and Audio
pyttsx3==2.98
gtts==2.5.4
pygame==2.6.1
speech_recognition==3.12.0
soundfile==0.12.1
librosa==0.10.2

# Weather and External APIs
python-weather==2.0.3
geopy==2.4.1

# Database
SQLAlchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10

# Utilities
python-dotenv==1.1.1
celery==5.4.0
redis==5.2.0
schedule==1.2.2
cryptography==44.0.0
bcrypt==4.2.1
email-validator==2.2.0
phonenumbers==8.13.50
pandas==2.2.3
numpy==2.1.3

# Development and Testing
pytest==8.3.4
pytest-flask==1.3.0
faker==33.1.0
black==24.10.0
flake8==7.1.1
pre-commit==4.0.1

# Production
gunicorn==23.0.0
supervisor==4.2.5

# Monitoring and Logging
sentry-sdk[flask]==2.19.2
prometheus-flask-exporter==0.23.1